将Health类型改为了float，为AI添加了低生命值逃离玩家寻找掩蔽的行为。

+++ Source/CoopGame/Private/Components/SHealthComponent.cpp
@@ -73,7 +73,7 @@ void USHealthComponent::HandleTakeAnyDamage(AActor* DamagedActor, float Damage, 
 
 }
 
-void USHealthComponent::Heal(int32 HealAmount)
+void USHealthComponent::Heal(float HealAmount)
 {
 	if (HealAmount <= 0 || Health <= 0)
 	{
